refactor: lens-debate — Server heartbeat reaper placed in kolu, not in @kolu/surface-app/server beside its sibling gateStaleSocket
Moved startWsHeartbeat/heartbeatSweep (and their tests) from kolu's packages/server into @kolu/surface-app/server, typed over a structural HeartbeatableSocket so surface-app keeps its no-ws-dependency stance and drishti inherits the reaper.

Agreed by the lowy ⇄ hickey lens debate (finding lowy-1, raised by lowy). Not pushed or merged.
